civicconnect-ai/requirements.txt
flask
python-dotenv
google-generativeai
Flask-Cors
Pillow
gunicorn